Created/moved test files into `playback-tree` directory to mirror what's in the src tree. Also added test coverage for `CommandNode` where promises resolve before execution or evaluation happens and return an error message.
