+++ packages/selenium-ide/src/neo/__test__/playback/playback-tree/syntax-validation.spec.js
@@ -15,8 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import { validateControlFlowSyntax } from "../../playback/playback-tree/syntax-validation";
-import Command, { ControlFlowCommandNames } from "../../models/Command";
+import { validateControlFlowSyntax } from "../../../playback/playback-tree/syntax-validation";
+import Command, { ControlFlowCommandNames } from "../../../models/Command";
 
 function createCommand(name) {
   return new Command(null, name, "", "");
